requirements.txt
altgraph==0.17.4
asttokens==2.4.1
blinker==1.9.0
buildozer==1.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
docutils==0.21.2
executing==2.0.1
filelock==3.13.1
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.1
jupyter_core==5.7.2
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.4
pefile==2023.2.7
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pygame==2.5.2
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pypiwin32==223
pyserial==3.5
python-dateutil==2.9.0.post0
pytube==15.0.0
pywin32==306
pywin32-ctypes==0.2.3
pyzmq==26.0.3
requests==2.32.3
setuptools==69.0.2
sh==2.1.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
urllib3==2.2.3
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==3.1.3
yt-dlp==2024.11.4
